requirements.txt
Package         Version
--------------- -------
-ip             20.2.4
cycler          0.10.0
decorator       4.4.2
kiwisolver      1.3.1
matplotlib      3.3.4
networkx        2.5
numpy           1.20.1
Pillow          8.1.2
pip             20.3.3
pyparsing       2.4.7
python-dateutil 2.8.1
setuptools      40.8.0
six             1.15.0
